Clusters of particles

Defines various cluster geometries and exports in a format suitable for TERMS

Helix of particles

Chain of particles

Core-satellite cluster of spheres

cluster_positions(
  N = 5,
  cl_fun = cluster_chain,
  radius = 50,
  label = "Au",
  ...,
  out = "",
  digits = 8
)

cluster_helix(
  N = 5,
  a = 10,
  b = 10,
  c = 50,
  R0 = 100,
  pitch = 200,
  delta = pi/5,
  delta0 = 0,
  hand = 1,
  ...
)

cluster_chain(
  N = 5,
  pitch = 500,
  a = 50,
  b = 30,
  c = b,
  rot = rotation_euler_passive(0, 0, 0)
)

cluster_satellite(
  N = 30,
  Rcore = 30,
  Rsat = 4,
  gap = 0.1,
  position = c("fibonacci", "landings", "hc"),
  exclusion = 10,
  ...
)

Arguments

N
number of particles

cl_fun
cluster function

radius
particle radius

label
particle material label

...
extra arguments passed to cl_fun

out
filename

digits
accuracy

a
semi-axis

b
semi-axis

c
semi-axis

R0
helix radius

pitch
chain pitch

delta
helix angle step

delta0
helix start angle

hand
helix handedness

rot
rotation matrix applied to each particle

Rcore
satellite radius

gap
gap distance

position
positioning on the sphere

exclusion
minimum exclusion distance for hc positions

Value

returns scatterers positions and sizes for an input file

Functions

cluster_positions: write cluster positions to input file

cluster_helix: helical cluster

cluster_chain: linear chain cluster

cluster_satellite: core-satellite cluster

Examples

cluster_positions()
#> Au -1000 0 0 50
#> Au -500 0 0 50
#> Au 0 0 0 50
#> Au 500 0 0 50
#> Au 1000 0 0 50
cluster_helix()
#> $positions
#>   [,1]      [,2]     [,3]      [,4]      [,5]
#> x  100  80.90170 30.90170 -30.90170 -80.90170
#> y    0  58.77853 95.10565  95.10565  58.77853
#> z  -40 -20.00000  0.00000  20.00000  40.00000
#> 
#> $sizes
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]   10   10   10   10   10
#> [2,]   10   10   10   10   10
#> [3,]   50   50   50   50   50
#> 
#> $angles
#>           [,1]     [,2]     [,3]      [,4]      [,5]
#> phi   1.570796 2.199115 2.827433 -2.827433 -2.199115
#> theta 1.262627 1.262627 1.262627  1.262627  1.262627
#> psi   0.000000 0.000000 0.000000  0.000000  0.000000
#> 
#> $R0
#> [1] 100
#> 
#> attr(,"class")
#> [1] "cluster"
cluster_chain()
#> $positions
#>       [,1] [,2] [,3] [,4] [,5]
#> [1,] -1000 -500    0  500 1000
#> [2,]     0    0    0    0    0
#> [3,]     0    0    0    0    0
#> 
#> $sizes
#>   [,1] [,2] [,3] [,4] [,5]
#> a   50   50   50   50   50
#> b   30   30   30   30   30
#> c   30   30   30   30   30
#> 
#> $angles
#>       [,1] [,2] [,3] [,4] [,5]
#> phi      0    0    0    0    0
#> theta    0    0    0    0    0
#> gamma    0    0    0    0    0
#> 
#> attr(,"class")
#> [1] "cluster"
cluster_satellite()
#> Error in sample_fibonacci(N) could not find function "sample_fibonacci"